techDebt(buRefactorTraining): Rename validation training class to trainingCsvValidator
