+++ server/models/BulkImport/csv/trainingCSVValidator.js
@@ -1,6 +1,6 @@
 const BUDI = require('../BUDI').BUDI;
 const moment = require('moment');
-class Training {
+class TrainingCsvValidator {
   constructor(currentLine, lineNumber) {
     this._currentLine = currentLine;
     this._lineNumber = lineNumber;
@@ -108,7 +108,7 @@ class Training {
         worker: this._currentLine.UNIQUEWORKERID,
         name: this._currentLine.LOCALESTID,
         lineNumber: this._lineNumber,
-        errCode: Training.LOCALESTID_ERROR,
+        errCode: TrainingCsvValidator.LOCALESTID_ERROR,
         errType: 'LOCALESTID_ERROR',
         error: 'LOCALESTID has not been supplied',
         source: this._currentLine.LOCALESTID,
@@ -120,7 +120,7 @@ class Training {
         worker: this._currentLine.UNIQUEWORKERID,
         name: this._currentLine.LOCALESTID,
         lineNumber: this._lineNumber,
-        errCode: Training.LOCALESTID_ERROR,
+        errCode: TrainingCsvValidator.LOCALESTID_ERROR,
         errType: 'LOCALESTID_ERROR',
         error: `LOCALESTID is longer than ${MAX_LENGTH} characters`,
         source: this._currentLine.LOCALESTID,
@@ -142,7 +142,7 @@ class Training {
         worker: this._currentLine.UNIQUEWORKERID,
         name: this._currentLine.LOCALESTID,
         lineNumber: this._lineNumber,
-        errCode: Training.UNIQUE_WORKER_ID_ERROR,
+        errCode: TrainingCsvValidator.UNIQUE_WORKER_ID_ERROR,
         errType: 'UNIQUE_WORKER_ID_ERROR',
         error: 'UNIQUEWORKERID has not been supplied',
         source: this._currentLine.UNIQUEWORKERID,
@@ -154,7 +154,7 @@ class Training {
         worker: this._currentLine.UNIQUEWORKERID,
         name: this._currentLine.LOCALESTID,
         lineNumber: this._lineNumber,
-        errCode: Training.UNIQUE_WORKER_ID_ERROR,
+        errCode: TrainingCsvValidator.UNIQUE_WORKER_ID_ERROR,
         errType: 'UNIQUE_WORKER_ID_ERROR',
         error: `UNIQUEWORKERID is longer than ${MAX_LENGTH} characters`,
         source: this._currentLine.UNIQUEWORKERID,
@@ -179,7 +179,7 @@ class Training {
           worker: this._currentLine.UNIQUEWORKERID,
           name: this._currentLine.LOCALESTID,
           lineNumber: this._lineNumber,
-          errCode: Training.DATE_COMPLETED_ERROR,
+          errCode: TrainingCsvValidator.DATE_COMPLETED_ERROR,
           errType: 'DATE_COMPLETED_ERROR',
           error: 'DATECOMPLETED is incorrectly formatted',
           source: this._currentLine.DATECOMPLETED,
@@ -191,7 +191,7 @@ class Training {
           worker: this._currentLine.UNIQUEWORKERID,
           name: this._currentLine.LOCALESTID,
           lineNumber: this._lineNumber,
-          errCode: Training.DATE_COMPLETED_ERROR,
+          errCode: TrainingCsvValidator.DATE_COMPLETED_ERROR,
           errType: 'DATE_COMPLETED_ERROR',
           error: 'DATECOMPLETED is invalid',
           source: this._currentLine.DATECOMPLETED,
@@ -203,7 +203,7 @@ class Training {
           worker: this._currentLine.UNIQUEWORKERID,
           name: this._currentLine.LOCALESTID,
           lineNumber: this._lineNumber,
-          errCode: Training.DATE_COMPLETED_ERROR,
+          errCode: TrainingCsvValidator.DATE_COMPLETED_ERROR,
           errType: 'DATE_COMPLETED_ERROR',
           error: 'DATECOMPLETED is in the future',
           source: this._currentLine.DATECOMPLETED,
@@ -233,7 +233,7 @@ class Training {
           worker: this._currentLine.UNIQUEWORKERID,
           name: this._currentLine.LOCALESTID,
           lineNumber: this._lineNumber,
-          errCode: Training.EXPIRY_DATE_ERROR,
+          errCode: TrainingCsvValidator.EXPIRY_DATE_ERROR,
           errType: 'EXPIRY_DATE_ERROR',
           error: 'EXPIRYDATE is incorrectly formatted',
           source: this._currentLine.EXPIRYDATE,
@@ -245,7 +245,7 @@ class Training {
           worker: this._currentLine.UNIQUEWORKERID,
           name: this._currentLine.LOCALESTID,
           lineNumber: this._lineNumber,
-          errCode: Training.EXPIRY_DATE_ERROR,
+          errCode: TrainingCsvValidator.EXPIRY_DATE_ERROR,
           errType: 'EXPIRY_DATE_ERROR',
           error: 'EXPIRYDATE is invalid',
           source: this._currentLine.EXPIRYDATE,
@@ -257,7 +257,7 @@ class Training {
           worker: this._currentLine.UNIQUEWORKERID,
           name: this._currentLine.LOCALESTID,
           lineNumber: this._lineNumber,
-          errCode: Training.EXPIRY_DATE_ERROR,
+          errCode: TrainingCsvValidator.EXPIRY_DATE_ERROR,
           errType: 'EXPIRY_DATE_ERROR',
           error: 'EXPIRYDATE must be after DATECOMPLETED',
           source: this._currentLine.EXPIRYDATE,
@@ -282,7 +282,7 @@ class Training {
         worker: this._currentLine.UNIQUEWORKERID,
         name: this._currentLine.LOCALESTID,
         lineNumber: this._lineNumber,
-        errCode: Training.DESCRIPTION_ERROR,
+        errCode: TrainingCsvValidator.DESCRIPTION_ERROR,
         errType: 'DESCRIPTION_ERROR',
         error: 'DESCRIPTION has not been supplied',
         source: this._currentLine.DESCRIPTION,
@@ -294,7 +294,7 @@ class Training {
         worker: this._currentLine.UNIQUEWORKERID,
         name: this._currentLine.LOCALESTID,
         lineNumber: this._lineNumber,
-        errCode: Training.DESCRIPTION_ERROR,
+        errCode: TrainingCsvValidator.DESCRIPTION_ERROR,
         errType: 'DESCRIPTION_ERROR',
         error: `DESCRIPTION is longer than ${MAX_LENGTH} characters`,
         source: this._currentLine.DESCRIPTION,
@@ -315,7 +315,7 @@ class Training {
         worker: this._currentLine.UNIQUEWORKERID,
         name: this._currentLine.LOCALESTID,
         lineNumber: this._lineNumber,
-        errCode: Training.CATEGORY_ERROR,
+        errCode: TrainingCsvValidator.CATEGORY_ERROR,
         errType: 'CATEGORY_ERROR',
         error: 'CATEGORY has not been supplied',
         source: this._currentLine.CATEGORY,
@@ -337,7 +337,7 @@ class Training {
           worker: this._currentLine.UNIQUEWORKERID,
           name: this._currentLine.LOCALESTID,
           lineNumber: this._lineNumber,
-          errCode: Training.ACCREDITED_ERROR,
+          errCode: TrainingCsvValidator.ACCREDITED_ERROR,
           errType: 'ACCREDITED_ERROR',
           error: 'ACCREDITED is invalid',
           source: this._currentLine.ACCREDITED,
@@ -371,7 +371,7 @@ class Training {
           worker: this._currentLine.UNIQUEWORKERID,
           name: this._currentLine.LOCALESTID,
           lineNumber: this._lineNumber,
-          errCode: Training.CATEGORY_ERROR,
+          errCode: TrainingCsvValidator.CATEGORY_ERROR,
           errType: 'CATEGORY_ERROR',
           error: 'CATEGORY has not been supplied',
           source: this._currentLine.CATEGORY,
@@ -393,7 +393,7 @@ class Training {
           worker: this._currentLine.UNIQUEWORKERID,
           name: this._currentLine.LOCALESTID,
           lineNumber: this._lineNumber,
-          errCode: Training.NOTES_ERROR,
+          errCode: TrainingCsvValidator.NOTES_ERROR,
           errType: 'NOTES_ERROR',
           error: `NOTES is longer than ${MAX_LENGTH} characters`,
           source: this._currentLine.NOTES,
@@ -407,7 +407,7 @@ class Training {
     }
   }
 
-  // returns true on success, false is any attribute of Training fails
+  // returns true on success, false is any attribute of TrainingCsvValidator fails
   validate() {
     let status = true;
     status = !this._validateLocaleStId() ? false : status;
@@ -470,4 +470,4 @@ class Training {
   }
 }
 
-module.exports.Training = Training;
+module.exports.TrainingCsvValidator = TrainingCsvValidator;
